requirements.txt
Django==4.1.7
djangorestframework==3.14.0
mysqlclient==2.1.1
django-model-utils==4.3.1
django-apscheduler==0.6.2
redis==4.5.1

# APScheduler==3.10.1
# asgiref==3.6.0
# pytz==2022.7.1
# pytz-deprecation-shim==0.1.0.post0
# six==1.16.0
# sqlparse==0.4.3
# tzdata==2022.7
# tzlocal==4.2
